Factor out ContributionUpdateAmountForm
